backend/requirements.txt
Flask==3.1.2
flask-cors==6.0.2
Flask-JWT-Extended==4.7.1
flask-marshmallow==1.3.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
marshmallow==4.1.2
python-dotenv==1.2.1
Werkzeug==3.1.4

